requirements.txt
numpy
pandas
matplotlib
kaggle
sofar
imageio
opencv-python
pyav
seaborn
pysofaconventions
torchsummary
tensorboard